+++ src/compressor.cpp
@@ -6,6 +6,7 @@
 #include <compressor.h>
 
 #include <pubkey.h>
+#include <script/script.h>
 #include <script/standard.h>
 
 /*
