+++ src/script/script.cpp
@@ -5,9 +5,12 @@
 
 #include <script/script.h>
 
+#include <hash.h>
 #include <util/strencodings.h>
 
 #include <string>
+
+CScriptID::CScriptID(const CScript& in) : BaseHash(Hash160(in)) {}
 
 std::string GetOpName(opcodetype opcode)
 {
