+++ src/script/sign.cpp
@@ -11,6 +11,7 @@
 #include <primitives/transaction.h>
 #include <script/keyorigin.h>
 #include <script/miniscript.h>
+#include <script/script.h>
 #include <script/signingprovider.h>
 #include <script/standard.h>
 #include <uint256.h>
