+++ src/test/fuzz/integer.cpp
@@ -19,6 +19,7 @@
 #include <pow.h>
 #include <protocol.h>
 #include <pubkey.h>
+#include <script/script.h>
 #include <script/standard.h>
 #include <serialize.h>
 #include <streams.h>
