+++ src/wallet/scriptpubkeyman.cpp
@@ -7,6 +7,7 @@
 #include <logging.h>
 #include <outputtype.h>
 #include <script/descriptor.h>
+#include <script/script.h>
 #include <script/sign.h>
 #include <util/bip32.h>
 #include <util/strencodings.h>
