+++ src/wallet/scriptpubkeyman.h
@@ -8,6 +8,7 @@
 #include <logging.h>
 #include <psbt.h>
 #include <script/descriptor.h>
+#include <script/script.h>
 #include <script/signingprovider.h>
 #include <script/standard.h>
 #include <util/error.h>
