+++ src/wallet/spend.cpp
@@ -11,6 +11,7 @@
 #include <numeric>
 #include <policy/policy.h>
 #include <primitives/transaction.h>
+#include <script/script.h>
 #include <script/signingprovider.h>
 #include <util/check.h>
 #include <util/fees.h>
