+++ src/wallet/walletdb.cpp
@@ -8,6 +8,7 @@
 #include <common/system.h>
 #include <key_io.h>
 #include <protocol.h>
+#include <script/script.h>
 #include <serialize.h>
 #include <sync.h>
 #include <util/bip32.h>
